✅ Finalmente el modulo de Institution esta completo y con varias validaciones rigurosas. No quito que aun despues tenga que agregar algunos endpoints adicionales para el tema de las rutas de consulta publica

+++ src/institution/institution.service.ts
@@ -1,10 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { CreateInstitutionDto } from './dto/create-institution.dto';
 import { UpdateInstitutionDto } from './dto/update-institution.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Institution } from './entities/institution.entity';
 import { Repository } from 'typeorm';
 import { QueriesInstitutionDto } from './dto/queries-institution.dto';
+import { User } from 'src/user/entities/user.entity';
+import { ROLE } from 'src/common/enums/role.enum';
 
 @Injectable()
 export class InstitutionService {
@@ -16,27 +18,32 @@ export class InstitutionService {
     async create(
         createInstitutionDto: CreateInstitutionDto,
         query: QueriesInstitutionDto,
+        user: User,
     ) {
-        const created = await this.repository.save(createInstitutionDto);
+        let created = null;
+        if (user.role === ROLE.ADMIN)
+            created = await this.repository.save(createInstitutionDto);
+        else
+            created = await this.repository.save({
+                ...createInstitutionDto,
+                user: user,
+            });
+
         return this.repository.findOne({
-            where: {
-                id: created.id,
-            },
+            where: { id: created.id },
             relations: query.include,
         });
     }
 
-    findAll(query: QueriesInstitutionDto) {
+    findAll(query: QueriesInstitutionDto, user: User) {
+        if (user.role === ROLE.ADMIN) {
+            return this.repository.find({
+                relations: query.include,
+            });
+        }
+
         return this.repository.find({
-            relations: query.include,
-        });
-    }
-
-    findOne(id: number, query: QueriesInstitutionDto) {
-        return this.repository.findOne({
-            where: {
-                id,
-            },
+            where: { user: user },
             relations: query.include,
         });
     }
@@ -45,25 +52,47 @@ export class InstitutionService {
         id: number,
         updateInstitutionDto: UpdateInstitutionDto,
         query: QueriesInstitutionDto,
+        user: User,
     ) {
-        await this.repository.save({
-            id,
-            ...updateInstitutionDto,
-        });
-        return this.repository.findOne({
-            where: {
+        if (user.role === ROLE.ADMIN)
+            await this.repository.save({
                 id,
-            },
+                ...updateInstitutionDto,
+                user: updateInstitutionDto.user, // Hay que definir explícitamente la relación que no se mapea automáticamente
+            });
+        else
+            await this.repository.save({
+                id,
+                ...updateInstitutionDto,
+                user: user,
+            });
+
+        return this.repository.findOne({
+            where: { id },
             relations: query.include,
         });
     }
 
-    async remove(id: number) {
-        const result = await this.repository.delete(id);
+    async remove(id: number, user: User) {
+        let result = null;
+        if (user.role === ROLE.ADMIN) result = await this.repository.delete(id);
+        else {
+            const institution = await this.repository.findOne({
+                where: { id },
+                relations: ['user'],
+            });
+
+            if (institution.user.id === user.id)
+                result = await this.repository.delete(id);
+            else
+                throw new ForbiddenException(
+                    'Solo puedes eliminar tus instituciones',
+                );
+        }
+
         return {
-            message: `Institution with id ${id} has been deleted`,
             ...result,
-            id,
+            message: `Institución con id ${id} ha sido eliminada`,
         };
     }
 }
